ARTEMIS-340 Add JDBC Storage Manager
