+++ artemis-core-client/src/main/java/org/apache/activemq/artemis/api/config/ActiveMQDefaultConfiguration.java
@@ -396,6 +396,18 @@ public final class ActiveMQDefaultConfiguration {
    // How often the reaper will be run to check for timed out group bindings. Only valid for LOCAL handlers
    private static long DEFAULT_GROUPING_HANDLER_REAPER_PERIOD = 30000;
 
+   // Which store type to use, options are FILE or DATABASE, FILE is default.
+   private static String DEFAULT_STORE_TYPE = "FILE";
+
+   // Default database url.  Derby database is used by default.
+   private static String DEFAULT_DATABASE_URL = "jdbc:derby:data/derby;create=true";
+
+   // Default message table name, used with Database storage type
+   private static String DEFAULT_MESSAGE_TABLE_NAME = "MESSAGES";
+
+   // Default bindings table name, used with Database storage type
+   private static String DEFAULT_BINDINGS_TABLE_NAME = "BINDINGS";
+
    /**
     * If true then the ActiveMQ Artemis Server will make use of any Protocol Managers that are in available on the classpath. If false then only the core protocol will be available, unless in Embedded mode where users can inject their own Protocol Managers.
     */
@@ -1052,4 +1064,28 @@ public final class ActiveMQDefaultConfiguration {
       return DEFAULT_GROUPING_HANDLER_REAPER_PERIOD;
    }
 
+   /**
+    * The default storage type.  Options are FILE and DATABASE.
+    */
+   public static String getDefaultStoreType() {
+      return DEFAULT_STORE_TYPE;
+   }
+
+   /**
+    * The default database URL, used with DATABASE store type.
+    */
+   public static String getDefaultDatabaseUrl() {
+      return DEFAULT_DATABASE_URL;
+   }
+
+   /**
+    * The default Message Journal table name, used with DATABASE store.
+    */
+   public static String getDefaultMessageTableName() {
+      return DEFAULT_MESSAGE_TABLE_NAME;
+   }
+
+   public static String getDefaultBindingsTableName() {
+      return DEFAULT_BINDINGS_TABLE_NAME;
+   }
 }
