+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/config/StoreConfiguration.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements. See the NOTICE file distributed with
  * this work for additional information regarding copyright ownership.
@@ -6,7 +6,7 @@
  * (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,16 +14,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.activemq.artemis.core.config;
 
-package org.apache.activemq.artemis.core.security;
+import java.io.Serializable;
 
-import org.apache.activemq.artemis.spi.core.protocol.RemotingConnection;
+public interface StoreConfiguration extends Serializable {
 
-public interface SecurityAuth {
+   public enum StoreType {
+      FILE,
+      DATABASE
+   }
 
-   String getUsername();
-
-   String getPassword();
-
-   RemotingConnection getRemotingConnection();
+   StoreType getStoreType();
 }
