+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/persistence/impl/journal/AddMessageRecord.java
@@ -26,11 +26,9 @@ public final class AddMessageRecord {
 
    final ServerMessage message;
 
-   // mtaylor (Added to compile)
-   public long scheduledDeliveryTime;
+   private long scheduledDeliveryTime;
 
-   // mtaylor (Added to compile)
-   public int deliveryCount;
+   private int deliveryCount;
 
    public ServerMessage getMessage() {
       return message;
@@ -44,4 +42,11 @@ public final class AddMessageRecord {
       return deliveryCount;
    }
 
+   public void setScheduledDeliveryTime(long scheduledDeliveryTime) {
+      this.scheduledDeliveryTime = scheduledDeliveryTime;
+   }
+
+   public void setDeliveryCount(int deliveryCount) {
+      this.deliveryCount = deliveryCount;
+   }
 }
