+++ pom.xml
@@ -47,6 +47,7 @@
       <module>artemis-ra</module>
       <module>artemis-rest</module>
       <module>artemis-service-extensions</module>
+      <module>artemis-jdbc-store</module>
       <module>artemis-maven-plugin</module>
       <module>artemis-server-osgi</module>
       <module>integration/activemq-spring-integration</module>
@@ -82,6 +83,7 @@
       <resteasy.version>3.0.13.Final</resteasy.version>
       <proton.version>0.10</proton.version>
       <fuse.mqtt.client.version>1.10</fuse.mqtt.client.version>
+      <apache.derby.version>10.11.1.1</apache.derby.version>
       <skipUnitTests>true</skipUnitTests>
       <skipJmsTests>true</skipJmsTests>
       <skipExtraTests>true</skipExtraTests>
@@ -201,6 +203,11 @@
             <artifactId>mqtt-client</artifactId>
             <version>${fuse.mqtt.client.version}</version>
             <!-- Apache v2.0 License -->
+         </dependency>
+         <dependency>
+            <groupId>org.apache.derby</groupId>
+            <artifactId>derby</artifactId>
+            <version>${apache.derby.version}</version>
          </dependency>
          <dependency>
             <groupId>org.eclipse.paho</groupId>
@@ -588,10 +595,12 @@
             <module>artemis-native</module>
             <module>artemis-protocols</module>
             <module>artemis-journal</module>
+            <module>artemis-jdbc-store</module>
             <module>artemis-ra</module>
             <module>artemis-rest</module>
             <module>artemis-service-extensions</module>
             <module>artemis-maven-plugin</module>
+            <module>artemis-jdbc-store</module>
             <module>integration/activemq-spring-integration</module>
             <module>integration/activemq-aerogear-integration</module>
             <module>integration/activemq-vertx-integration</module>
@@ -623,6 +632,7 @@
             <module>artemis-jms-server</module>
             <module>artemis-native</module>
             <module>artemis-journal</module>
+            <module>artemis-jdbc-store</module>
             <module>artemis-ra</module>
             <module>artemis-rest</module>
             <module>artemis-service-extensions</module>
@@ -681,6 +691,7 @@
             <module>artemis-jms-server</module>
             <module>artemis-native</module>
             <module>artemis-journal</module>
+            <module>artemis-jdbc-store</module>
             <module>artemis-ra</module>
             <module>artemis-rest</module>
             <module>artemis-service-extensions</module>
@@ -723,6 +734,7 @@
             <module>artemis-jms-server</module>
             <module>artemis-native</module>
             <module>artemis-journal</module>
+            <module>artemis-jdbc-store</module>
             <module>artemis-ra</module>
             <module>artemis-rest</module>
             <module>artemis-service-extensions</module>
@@ -757,6 +769,7 @@
             <module>artemis-jms-server</module>
             <module>artemis-native</module>
             <module>artemis-journal</module>
+            <module>artemis-jdbc-store</module>
             <module>artemis-ra</module>
             <module>artemis-rest</module>
             <module>artemis-service-extensions</module>
